+++ docs/changelog.md
@@ -8,6 +8,7 @@
 - `borax.calendars.festivals2`
   - 新增 `Festival.gets` 获取属性方法
   - 新增 `Festival.description` 属性，表示规范化中文描述
+  - 新增 `FestivalLibrary.extend_unique` ，支持以去重方式添加新节日
 - `borax.calendars.utils`
   - 新增 `ThreeNineUtils` 三伏九九天解析工具
 - 开发SOP